# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.LoginRequest:
    properties:
      password:
        example: password123
        type: string
      username:
        example: budi
        type: string
    required:
    - password
    - username
    type: object
  controllers.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
  controllers.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  controllers.RegisterRequest:
    properties:
      email:
        example: budi@example.com
        type: string
      name:
        example: Budiawan Bengi
        type: string
      password:
        example: password123
        minLength: 6
        type: string
      username:
        example: budi
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - name
    - password
    - username
    type: object
...
  models.RoleResponse:
    properties:
      assigned_at:
        type: string
      assigned_by:
        type: string
      description:
        type: string
      id:
        type: integer
      role:
        type: string
    type: object
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.RoleResponse'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact:
    email: support@livotech.com
    name: Saya
  description: Layanan backend manajemen pengguna yang komprehensif dengan autentikasi
    JWT dan kontrol akses berbasis role
  title: Livotech Backend Service
  version: "2.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Mengautentikasi pengguna dan mengembalikan token JWT
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Login user
      tags:
      - auth
  /api/auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user dengan menonaktifkan refresh token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: Memperbarui access token baru menggunakan token refresh
      parameters:
      - description: Refresh token request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Refresh access token
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Mendaftarkan akun pengguna baru
      parameters:
      - description: Registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Register user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Ketik "Bearer" diikuti oleh spasi dan token JWT.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"